feat: update middleware
