+++ src/app/login/components/InputForm.jsx
@@ -1,8 +1,8 @@
 const InputForm = ({ label, type, name, error }) => {
     return (
         <div className="flex flex-col w-full gap-2">
-            <label htmlFor={name} className="font-semibold capitalize">{label}</label>
-            <input type={type} id={name} name={name} className="bg-white/5 border rounded-md h-10 border-white/10 px-2 focus:outline-white/40" />
+            <label htmlFor={name} className="font-semibold capitalize text-sm">{label}</label>
+            <input type={type} id={name} name={name} className="bg-white/5 border rounded-md h-10 border-white/10 px-2 focus:outline-white/40 text-sm" />
             {
                 error ? <p className="text-sm text-white/60 ">{error}</p> : <div className="h-1"></div>
             }
