+++ src/app/login/components/SubmitButton.jsx
@@ -1,6 +1,6 @@
 const SubmitButton = ({children}) => {
     return (
-        <button className="bg-white rounded-lg h-10 w-full text-black/80 font-bold capitalize text-sm">
+        <button className="bg-white rounded-md h-10 w-full text-black/80 font-bold capitalize text-sm">
             {children}
         </button>
     );
