+++ src/middleware.js
@@ -11,7 +11,7 @@ export default async function middleware(req) {
  
   const cookie = (await cookies()).get('access_token')?.value
 
-  if (isProtectedRoute && !cookie) {
+  if ((isProtectedRoute || path.startsWith(protectedRoutes[0])) && !cookie) {
     return NextResponse.redirect(new URL('/login', req.nextUrl))
   }
