+++ tsconfig.json
@@ -28,7 +28,7 @@
     ".next/types/**/*.ts",
     "**/*.ts",
     "**/*.tsx"
-, "src/app/login/page.js", "src/app/login/components/InputForm.jsx", "src/app/login/components/SubmitButton.jsx", "src/middleware.js", "src/app/api/session/route.js", "src/app/dashboard/proctored_users/page.tsx", "src/app/dashboard/sessions/[userId]/page.js", "src/app/dashboard/settings/page.tsx", "src/app/dashboard/room/ui/SideBarLog.jsx", "src/app/dashboard/room/lib/flagtype.js", "src/app/dashboard/room/[roomId]/[socketId]/page.tsx", "src/app/dashboard/proctored_users/[userId]/analytics/[token]/page.tsx"  ],
+, "src/app/login/page.tsx", "src/app/login/components/InputForm.jsx", "src/app/login/components/SubmitButton.jsx", "src/middleware.js", "src/app/api/session/route.js", "src/app/dashboard/proctored_users/page.tsx", "src/app/dashboard/sessions/[userId]/page.js", "src/app/dashboard/settings/page.tsx", "src/app/dashboard/room/ui/SideBarLog.jsx", "src/app/dashboard/room/lib/flagtype.js", "src/app/dashboard/room/[roomId]/[socketId]/page.tsx", "src/app/dashboard/proctored_users/[userId]/analytics/[token]/page.tsx"  ],
   "exclude": [
     "node_modules"
   ]
